backend/requirements.txt
# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Options Pricing and Quantitative Finance
QuantLib>=1.30
yfinance>=0.2.28

# Tastyworks Trading API (unofficial)
tastytrade>=6.0

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=2.0.0

# Time Series Database
# Note: TimescaleDB is a PostgreSQL extension, installed separately

# Data Storage
arctic>=1.79.0  # For tick data storage
pyarrow>=12.0.0

# Web Framework for Dashboard
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
fastapi>=0.104.0
uvicorn>=0.24.0

# AI/ML APIs
openai>=1.3.0

# Async Operations
asyncio>=3.4.3
aiohttp>=3.8.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development Tools
python-dotenv>=1.0.0
black>=23.7.0
pylint>=2.17.0
mypy>=1.5.0

# Visualization (for monitoring dashboard)
plotly>=5.15.0
dash>=2.11.0

# Additional ML Libraries (optional but useful)
xgboost>=1.7.0
lightgbm>=4.0.0

# Risk Management
riskfolio-lib>=4.0.0
cvxpy>=1.3.0